requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
cffi==1.17.0
cryptography==43.0.0
Django==4.2
djangorestframework==3.15.2
numpy==2.1.0
pandas==2.2.2
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
